requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-unfold==0.47.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-magic==0.4.27
sqlparse==0.4.4
tzdata==2023.3
